\docType{data}
\name{sipp84fp_dct}
\alias{sipp84fp_dct}
\title{Survey of Income and Program Participation (SIPP) - 1984 Panel}
\source{
  \url{http://thedataweb.rm.census.gov/ftp/sipp_ftp.html#sipp84}
}
\description{
  Dictionary file for the full 1984 panel for NBER's Survey
  of Income and Program Participation (SIPP).
}
\keyword{datasets}

