\docType{package}
\name{StataDCTutils-package}
\alias{StataDCTutils}
\alias{StataDCTutils-package}
\title{Parses Stata dictionary files for further use with R}
\description{
  Utilities to make reading fixed-width-format datasets
  distributed as a \code{.dat}/\code{.dct}/\code{.do} set
  of Stata files more convenient with R.
}
\details{
  \tabular{ll}{ Package: \tab StataDCTutils\cr Type: \tab
  Package\cr Version: \tab 1.0\cr Date: \tab 2013-01-16\cr
  License: \tab CC-SA\cr }
}
\author{
  Ananda Mahto

  Maintainer: Ananda Mahto <mrdwab@gmail.com>
}
\references{
  Initial versions of this function can be found at
  \url{http://stackoverflow.com/questions/14224321/reading-dat-and-dct-directly-from-r}
}
\keyword{package}

