\docType{data}
\name{data66_dat}
\alias{data66_dat}
\title{Example dataset to be converted to a csv file}
\source{
  Obtained from
  \url{http://faculty-staff.ou.edu/L/Carlos.E.Lamarche-1/ec5243/data.html}
}
\description{
  This is a sample dataset useful to demonstrate conversion
  to csv using \code{\link{csvkit.fwf2csv}} after parsing
  the dictionary file (using \code{\link{dct.parser}}) and
  creating a csvkit schema file (using
  \code{\link{csvkit.schema}}).
}
\keyword{datasets}

