\docType{data}
\name{data66_dct}
\alias{data66_dct}
\title{Example dictionary file accompanying \code{\link{data66_dat}}}
\source{
  Obtained from
  \url{http://faculty-staff.ou.edu/L/Carlos.E.Lamarche-1/ec5243/data.html}
}
\description{
  Example dictionary file accompanying
  \code{\link{data66_dat}} to be parsed using
  \code{\link{dct.parser}} and then possibly using other
  tools from the \code{StataDCTutils} package.
}
\keyword{datasets}

