\name{sipp84fp_dct}
\alias{sipp84fp_dct}
\docType{data}
\title{
Survey of Income and Program Participation (SIPP) - 1984 Panel
}
\description{
Dictionary file for the full 1984 panel for NBER's Survey of Income and Program Participation (SIPP).
}
\usage{
data(sampleDctData)
## sipp84fp_dct is a character vector
## Write the dictionary to a text file in your working directory
writeLines(sipp84fp_dct, "sipp84fp.dct")
}
\source{
\url{http://thedataweb.rm.census.gov/ftp/sipp_ftp.html#sipp84}
}
\keyword{datasets}
